JIT: When making memory writable, include the exception handler
makeWritable() rounds the memory down to the next page boundary. Usually
we include the exception handler this way, unless the offset from the
page boundary is less than the exception handler size. Make it explicit
that we do want the exception handler to be writable, too.

Fixes: QTBUG-89513
Pick-to: 5.15 6.0
Change-Id: I2fb8fb0e1dcc3450b036924463dc1b40d2020c46
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>

Cherry-picked from dev to adsk-contrib-maya-v5.15.2 by Keith Kyzivat 20221128
This fix present in v5.15.3