only safe to call QWindow::setCursor() when the cursorItem is really this
When QQuickItem::setCursor() is called from webengine, there is no
guarantee the offscreen window's cursor item is this. When using
QQuickWidget, the cursor item can be out of date because the mouse
cursor can move out of the quick widget.

We need to call updateCursor() first to update the cursor item. Then,
double check the cursor item is really this. It's only safe to
call the render window's setCursor() when the cursor item is really
this.