MAYA-139846: Crash on reparenting (#7)
Upon reparenting widgets containing RHI widgets in the hierarchy, such as QuickWidget, the RHI can be different from what's provided by QWidgetPrivate::rhi. In such a case, it's not only important to reset the cached rhi to null but also to invalidate the render control so it's re-initialized at a later point. Otherwise, the render control will hold onto the dead rhi, resulting in a crash. The idea of invaliding the render control is also used in other areas of the code in this file.